refactor(cli): ship the guren add blueprint scaffolds as real sources
Move every fully static file the guren add blueprints emit (admin, oauth,
cache, events, mail, queue, notifications, storage, broadcasting,
schedule) out of template literals in blueprints.ts into real sources
under templates/scaffold/<blueprint>/, loaded via loadScaffoldTemplate().
A string-emitted file is a file no tsconfig covers; these now typecheck
in the tsconfig.templates*.json projects like the auth and attachments
scaffolds already do.

Flag-dependent output stays inline: the admin controller and routes file
vary with --public, so they remain builder-emitted.

mail gets its own tsconfig.templates-mail.json chained into
typecheck:templates — rootDirs merges every listed scaffold into one
virtual root, and auth already ships app/Providers/MailProvider.ts at
the same relative path.

The events/queue/broadcasting providers import files their blueprints
generate through make:event/listener/job/channel; those renders live as
fixtures under tests/fixtures/scaffold-typecheck/<blueprint>/ and are
pinned whole-file (toBe) against their builders in
scaffold-output.test.ts. A new gate there also runs every remaining
blueprint and asserts each shipped template lands byte-identical, so a
template no blueprint writes cannot ship as dead weight.

Blueprint output is byte-identical to what the literals emitted, so no
changeset is needed.
