+++ packages/cli/templates/scaffold/cache/app/Providers/CacheProvider.ts
@@ -1,0 +1,12 @@
+import { ServiceProvider, createCacheManager } from '@guren/core'
+
+export default class CacheProvider extends ServiceProvider {
+  register(): void {
+    this.container.singleton('cache', () => createCacheManager({
+      default: 'memory',
+      stores: {
+        memory: { driver: 'memory' },
+      },
+    }))
+  }
+}
