+++ packages/cli/templates/scaffold/cache/app/Services/ApplicationCache.ts
@@ -1,0 +1,9 @@
+import type { CacheManager } from '@guren/core'
+
+export class ApplicationCache {
+  constructor(private readonly cache: CacheManager) {}
+
+  async rememberVersion(): Promise<string> {
+    return this.cache.store().rememberForever('app:version', async () => 'vNext')
+  }
+}
