+++ packages/cli/templates/scaffold/schedule/app/Console/Kernel.ts
@@ -1,0 +1,16 @@
+import { Schedule } from '@guren/core'
+
+export function scheduleTasksKernel(): Schedule {
+  const schedule = new Schedule()
+
+  schedule
+    .call(async () => {
+      console.log('[Schedule] Running heartbeat task')
+    })
+    .hourly()
+    .name('app-heartbeat')
+
+  return schedule
+}
+
+export default scheduleTasksKernel
