+++ packages/cli/tsconfig.templates-mail.json
@@ -1,0 +1,29 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "paths": {
+      "@guren/core": [
+        "../core/src/index.ts"
+      ],
+      "@guren/core/*": [
+        "../core/src/*"
+      ],
+      "@guren/server": [
+        "../server/src/index.ts"
+      ],
+      "@guren/server/*": [
+        "../server/src/*"
+      ],
+      "@guren/orm": [
+        "../orm/src/index.ts"
+      ],
+      "@guren/orm/*": [
+        "../orm/src/*"
+      ]
+    }
+  },
+  "include": [
+    "templates/scaffold/mail/**/*"
+  ],
+  "exclude": []
+}
